\name{chp2}
\alias{chp2}
\docType{data}
\title{
An example BED file.
}
\description{
This is an example BED File corresponding to the Ion AmpliSeq (TM) Cancer Hotspot Panel v2. It contains 207 regions that are frequently mutated in cancer.
}
\usage{data(chp2)}
\format{
A data frame with 207 observations on 6 variables. 
}
\references{
\url{http://tools.lifetechnologies.com/content/sfs/brochures/Ion-AmpliSeq-Cancer-Hotspot-Panel-Flyer.pdf}
}
\examples{
data(chp2, package="ionflows")
}
\keyword{datasets}
