\name{ionflows-package}
\alias{ionflows-package}
\alias{ionflows}
\docType{package}
\title{
Calculate the Number of Required Flows for Semiconductor Sequencing
}
\description{
This package implements two methods for calculation of the number of required flows for semiconductor sequencing: 1. Using a simulation, the number of flows can be calculated for a concrete list of amplicons. 2. An exact combinatorial model is evaluated to calculate the number of flows for a random ensemble of sequences.
}

\details{
\tabular{ll}{
Package: \tab ionflows\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-21\cr
License: \tab GPL-3 \cr
}

}
\author{
Michael Bockmayr <michael.bockmayr@charite.de>,

Jan Budczies <jan.budczies@charite.de>,

Institute of Pathology, Charité University Hospital,

Charitéplatz 1, 10117 Berlin, Germany
}
\references{
Budczies, J et al. \emph{Semiconductor sequencing: How many flows do you need?} Submitted. 
}

\keyword{package}

